TLS Finally working against an Android Client.
The client uses SSLSocketFactory and TrustManager classes to create a
secure context based on the trust of the REGISTRATION CA.pem file.

TLSv1.2 Handshake is negotiated at most, certificate chain file is well
structured and generated on C++. A TLS certificate is issued by the CA.

The Android client generates a CSR that is sent securely to the server,
which signs it and stores it on the certs/ directory.